// src/app/enum.ts
export enum Values {
    Ace = 1,
    Two = 2,
    Three = 3,
    Four = 4,
    Five = 5,
    Six = 6,
    Seven = 7,
    Eight = 8,
    Nine = 9,
    Ten = 10,
    Jack = 11,
    Queen = 12,
    King = 13
}

export enum Families {
    Hearts = 1,
    Spades = 2,
    Diamonds = 3,
    Clubs = 4
}

export enum Colors {
    Red = 1,
    Black = 2
}